requirements.txt
opencv-python-headless
pytesseract
pandas
numpy
Pillow
